Refactor init.sh
Some major changes: Remove confirm(), and use prompt() for everything.
Prompt() now also has extra options to set the default response.

Get() now only downloads the file if it doesn't already exist.

Added replace(), to download a prebuilt conf file and replace the local
conf file.

Moved IP address steps to before IP tables.

SSH is now a separate item.

The iptables script is now a systemd unit as well.

A number of file permissions are set just in case.

Splunk installation has been moved to the end of the script since it can
take a while and is non-essential.

+++ linux/init.sh
@@ -18,31 +18,52 @@ CCDC_DIR="/ccdc"
 DOWNLOAD_DIR="$CCDC_DIR/downloads"
 SCRIPT_DIR="$CCDC_DIR/scripts"
 
+# make directories and set current directory
 mkdir -p $CCDC_DIR
-cd $CCDC_DIR
 mkdir -p $DOWNLOAD_DIR
 mkdir -p $SCRIPT_DIR
+cd $CCDC_DIR
 
-confirm() {
-  read -r -p "$1 [Y/n]:" RES
-  [[ !("$RES" =~ ^([nN]).*$) ]]
-}
-
-get() {
-  mkdir -p $(dirname "$DOWNLOAD_DIR/$1")
-  wget "https://raw.githubusercontent.com/UWStout-CCDC/CCDC-scripts-2020/master/$1" -O "$DOWNLOAD_DIR/$1"
-}
-
+# if prompt <prompt> n; then; <cmds>; fi
+# Defaults to NO
+# if prompt <prompt> y; then; <cmds>; fi
+# Defaults to YES
 prompt() {
-  read -p "$1 [y/N]" ans
+  case "$2" in 
+    y) def="[y/N]" ;;
+    n) def="[Y/n]" ;;
+    *) echo "INVALID PARAMETER!!!!"; exit ;;
+  esac
+  read -p "$1 $def" ans
   case $ans in
     y|Y) true ;;
-    *) false ;;
+    n|N) false ;;
+    *) [[ "$def" != "[y/N]" ]] ;;
   esac
 }
 
-get linux/log_state.sh
-bash $DOWNLOAD_DIR/linux/log_state.sh
+# get <file>
+get() {
+  # only download if the file doesn't exist
+  if [[ ! -f "$DOWNLOAD_DIR/$1" ]]
+  then
+    mkdir -p $(dirname "$DOWNLOAD_DIR/$1")
+    wget "https://raw.githubusercontent.com/UWStout-CCDC/CCDC-scripts-2020/master/$1" -O "$DOWNLOAD_DIR/$1"
+  fi
+}
+
+# replace <dir> <file> <new file>
+replace() {
+  get $3
+  mkdir -p $CCDC_DIR/$(dirname $2)
+  cp $1/$2 $CCDC_DIR/$2.old
+  cp $DOWNLOAD_DIR/$3 $1/$2
+}
+
+# Grab script so it's guarnteed to be in /ccdc/scripts/linux
+get linux/init.sh
+
+get linux/log_state.sh && bash $DOWNLOAD_DIR/linux/log_state.sh
 
 #gets wanted username
 echo "What would you like the admin account to be named?"
@@ -72,6 +93,29 @@ echo "Set root password"
 passwd root
 
 bash $PASSWD_SH
+
+# Current IP address. We should assume this to be correct
+IP_ADDR=$(ip addr | grep -Po "inet \K172\.\d+\.\d+\.\d+")
+
+if prompt "Is $IP_ADDR the correct IP address?" y
+then
+else
+  read -p "Enter the correct IP address: " IP_ADDR
+fi
+
+# Force sets the ip address and dns server
+# TODO: Test this works on every server
+cat << EOF > /etc/network/interfaces
+auto lo
+iface lo inet loopback
+
+auto eth0
+iface eth0 inet static
+  address ${IP_ADDR}
+  netmask 255.255.255.0
+  gateway ${IP_ADDR%.*}.254
+  dns-nameserver 172.20.240.20 172.20.242.200 9.9.9.9
+EOF
 
 # Iptables
 IPTABLES_SCRIPT="$SCRIPT_DIR/iptables.sh"
@@ -110,73 +154,142 @@ iptables -t filter -A OUTPUT -p tcp --dport 80 -j ACCEPT
 iptables -t filter -A OUTPUT -p tcp --dport 443 -j ACCEPT
 
 # NTP (server time)
-iptables -t filter -a output -p udp --dport 123 -j accept
+iptables -t filter -a OUTPUT -p udp --dport 123 -j accept
 
 # Splunk
-iptables -t filter -a output -p tcp --dport 8000 -j accept
-iptables -t filter -a output -p tcp --dport 8089 -j accept
-iptables -t filter -a output -p tcp --dport 9997 -j accept
+iptables -t filter -a OUTPUT -p tcp --dport 8000 -j accept
+iptables -t filter -a OUTPUT -p tcp --dport 8089 -j accept
+iptables -t filter -a OUTPUT -p tcp --dport 9997 -j accept
+
+# SSH outbound
+iptables -A OUTPUT -p tcp --sport 22 -m conntrack --ctstate ESTABLISHED -j ACCEPT
 
 ######## OUTBOUND SERVICES ###############
 
-# SSH
-iptables -A INPUT -p tcp --dport 22 -m conntrack --ctstate NEW,ESTABLISHED -j ACCEPT
-iptables -A OUTPUT -p tcp --sport 22 -m conntrack --ctstate ESTABLISHED -j ACCEPT
-
 EOF
 
-confirm "HTTP(S) Server?" && cat <<EOF >> $IPTABLES_SCRIPT
-# HTTP/HTTPS (apache) iptables -t filter -A INPUT -p tcp --dport 80 -j ACCEPT
-iptables -t filter -A INPUT -p tcp --dport 443 -j ACCEPT
+if prompt "SSH Server?" n
+then
+  cat <<-EOF >> $IPTABLES_SCRIPT
+  # SSH
+  iptables -A INPUT -p tcp --dport 22 -m conntrack --ctstate NEW,ESTABLISHED -j ACCEPT
 
-EOF
+  EOF
+fi
 
-confirm "DNS Server?" && cat <<EOF >> $IPTABLES_SCRIPT
-# DNS (bind)
-iptables -t filter -A INPUT -p tcp --dport 53 -j ACCEPT
-iptables -t filter -A INPUT -p udp --dport 53 -j ACCEPT
+if prompt "HTTP(S) Server?" n
+then
+  cat <<-EOF >> $IPTABLES_SCRIPT
+  # HTTP/HTTPS (apache) iptables -t filter -A INPUT -p tcp --dport 80 -j ACCEPT
+  iptables -t filter -A INPUT -p tcp --dport 443 -j ACCEPT
 
-EOF
+  EOF
+fi
 
-confirm "MAIL Server?" && cat <<EOF >> $IPTABLES_SCRIPT
-# SMTP
-iptables -t filter -A OUTPUT -p tcp --dport 25 -j ACCEPT
-iptables -t filter -A INPUT -p tcp --dport 25 -j ACCEPT
 
-# POP3
-iptables -t filter -A OUTPUT -p tcp --dport 110 -j ACCEPT
-iptables -t filter -A INPUT -p tcp --dport 110 -j ACCEPT
+if prompt "DNS Server?" n
+then
+  cat <<-EOF >> $IPTABLES_SCRIPT
+  # DNS (bind)
+  iptables -t filter -A INPUT -p tcp --dport 53 -j ACCEPT
+  iptables -t filter -A INPUT -p udp --dport 53 -j ACCEPT
 
-# IMAP
-iptables -t filter -A OUTPUT -p tcp --dport 143 -j ACCEPT
-iptables -t filter -A INPUT -p tcp --dport 143 -j ACCEPT
+  EOF
+fi
 
-EOF
+if prompt "MAIL Server?" n
+then
+  cat <<-EOF >> $IPTABLES_SCRIPT
+  # SMTP
+  iptables -t filter -A OUTPUT -p tcp --dport 25 -j ACCEPT
+  iptables -t filter -A INPUT -p tcp --dport 25 -j ACCEPT
 
-confirm "NTP Server?" && cat <<EOF >> $IPTABLES_SCRIPT
-# NTP
-iptables -t filter -A INPUT -p tcp --dport 123 -j ACCEPT
-iptables -t filter -A INPUT -p udp --dport 123 -j ACCEPT
+  # POP3
+  iptables -t filter -A OUTPUT -p tcp --dport 110 -j ACCEPT
+  iptables -t filter -A INPUT -p tcp --dport 110 -j ACCEPT
 
-EOF
+  # IMAP
+  iptables -t filter -A OUTPUT -p tcp --dport 143 -j ACCEPT
+  iptables -t filter -A INPUT -p tcp --dport 143 -j ACCEPT
+
+  EOF
+fi
+
+if prompt "NTP Server?" n
+then
+  cat <<-EOF >> $IPTABLES_SCRIPT
+  # NTP
+  iptables -t filter -A INPUT -p tcp --dport 123 -j ACCEPT
+  iptables -t filter -A INPUT -p udp --dport 123 -j ACCEPT
+
+  EOF
+fi
 bash $IPTABLES_SCRIPT
 
+# Create systemd unit for the firewall
+mkdir -p /etc/systemd/system/
+cat << EOF > /etc/systemd/system/ccdc_firewall.service
+[Unit]
+Description=ZDSFirewall
+After=syslog.target network.target
+
+[Service]
+Type=oneshot
+ExecStart=$IPTABLES_SCRIPT
+ExecStop=/sbin/iptables -F
+RemainAfterExit=yes
+
+[Install]
+WantedBy=multi-user.target
+EOF
+
+# Set Legal banners
+replace /etc motd general/legal_banner.txt
+replace /etc issue general/legal_banner.txt
+replace /etc issue.net general/legal_banner.txt
+
+# Fix permissions (just in case)
+chown root:root /etc/group
+chmod a=r,u=rw /etc/group
+chown root:root /etc/sudoers
+chmod a=r,u=rw /etc/sudoers
+chown root:root /etc/passwd
+chmod a=r,u=rw /etc/passwd
+if [ $(getent group shadow) ]; then
+  chown root:shadow /etc/shadow
+else
+  chown root:root /etc/shadow
+fi
+chmod a=r,u=rw /etc/shadow
+
+
+# We might be able to get away with installing systemd on centos 6 to make every server the same
+
+# TODO: There are multiple ways to do NTP. We need to check what each server uses.
+#server 172.20.240.20
+# timedatectl status
+
 # SSH Server config
-get linux/sshd_config
-
-mkdir -p $CCDC_DIR/ssh/
-cp /etc/ssh/sshd_config $CCDC_DIR/ssh/sshd_config.old
-cp $DOWNLOAD_DIR/linux/sshd_config /etc/ssh/sshd_config
-
+replace /etc ssh/sshd_config linux/sshd_config
 # Disable all keys - sshd_config will set the server to check this file
 touch /ccdc/ssh/authorized_keys
+
+# !! DO LAST !! These will take a while
 
 # Restart service
 if type systemctl
 then
   systemctl restart sshd
+
+  # Disable other firewalls
+  systemctl disable --now firewalld
+  systemctl disable --now ufw
+
+  # Automatically apply IPTABLES_SCRIPT on boot
+  systemctl enable --now ccdc_firewall.service
 else
   service sshd restart
+  # On non-systemd systems, the firewall will need to be reapplied in another way
 fi
 
 if type yum
@@ -192,43 +305,7 @@ else
 fi
 
 # Splunk forwarder
-get linux/splunk.sh
-bash $DOWNLOAD_DIR/linux/splunk.sh 172.20.241.20
-
-get general/legal_banner.txt
-
-mkdir -p $CCDC_DIR/
-cp /etc/motd $CCDC_DIR/motd.old
-cp $DOWNLOAD_DIR/general/legal_banner.txt /etc/motd
-
-# Current IP address. We should assume this to be correct
-IP_ADDR=$(ip addr | grep -Po "inet \K172\.\d+\.\d+\.\d+")
-
-if prompt "Is $IP_ADDR the correct IP address?"
-then
-else
-  read -p "Enter the correct IP address: " IP_ADDR
-fi
-
-# Force sets the ip address and dns server
-# TODO: Test this works on every server
-cat << EOF > /etc/network/interfaces
-auto lo
-iface lo inet loopback
-
-auto eth0
-iface eth0 inet static
-  address ${IP_ADDR}
-  netmask 255.255.255.0
-  gateway ${IP_ADDR%.*}.254
-  dns-nameserver 172.20.240.20 172.20.242.200 9.9.9.9
-EOF
-
-# We might be able to get away with installing systemd on centos 6 to make every server the same
-
-# There are multiple ways to do NTP. We need to check what each server uses.
-#server 172.20.240.20
-# timedatectl status
+get linux/splunk.sh && bash $DOWNLOAD_DIR/linux/splunk.sh 172.20.241.20
 
 
-# Now restart the machine to guarntee all changes apply.
+echo "Now restart the machine to guarntee all changes apply"
